separate hosts package from main and reorganize package structure
add hosts, format and mackerelclient package
