+++ mackerelclient/testdata/mackerel-agent-no-base.conf
@@ -1,4 +1,4 @@
 pidfile = "./pid"
-root = "./test"
+root = "./testdata"
 verbose = false
 apikey = "ABCD123456"
