+++ mackerelclient/testdata/mackerel-agent.conf
@@ -1,5 +1,5 @@
 pidfile = "./pid"
-root = "./test"
+root = "./testdata"
 verbose = false
 apikey = "123456ABCD"
 apibase = "https://example.com/"
